// src/app/models/IXXE_WIP_OT.ts
export interface IXXE_WIP_OT {
  organizatioN_ID?: string | null;
  organizatioN_CODE?: string | null;
  wiP_ENTITY_NAME: string | null;
  starT_QUANTITY: string | null;
  quantitY_COMPLETED: string;
  aLTERNATE_BOM_DESIGNATOR?: string | null;
  primarY_ITEM_ID?: string | null;
  segmenT1: string | null;
  description?: string | null;
  transDetalles?: string;
}